[refactoring] refact name
